Move SQLite constants to constants.ts
