--- src/lib/sqlite-database-integration-release.ts
@@ -1,3 +1,0 @@
-export const SQLITE_DATABASE_INTEGRATION_VERSION = 'v2.1.17-alpha';
-
-export const SQLITE_DATABASE_INTEGRATION_RELEASE_URL = `https://github.com/Automattic/sqlite-database-integration/archive/refs/tags/${ SQLITE_DATABASE_INTEGRATION_VERSION }.zip`;
